+++ sdk/src/utils/ze_utils.h
@@ -338,8 +338,8 @@ inline std::vector<std::string> GetMetricList(zet_metric_group_handle_t group) {
   return name_list;
 }
 
-inline uint32_t GetMetricId(const std::vector<std::string>& metric_list,
-                            const std::string_view metric_name) {
+inline size_t GetMetricId(const std::vector<std::string>& metric_list,
+                          const std::string_view metric_name) {
   PTI_ASSERT(!metric_list.empty());
   PTI_ASSERT(!metric_name.empty());
